+++ packages/components/src/components/layout/table/table-cell/table-cell-styles.spec.ts
@@ -1,0 +1,7 @@
+import { getComponentCss } from './table-cell-styles';
+
+describe('getComponentCss()', () => {
+  it.each([[true], [false]])('should return correct css for multiline: %s', (multiline: boolean) => {
+    expect(getComponentCss(multiline)).toMatchSnapshot();
+  });
+});