+++ packages/components/src/components/layout/table/table-head-row/table-head-row-styles.spec.ts
@@ -1,0 +1,7 @@
+import { getComponentCss } from './table-head-row-styles';
+
+describe('getComponentCss()', () => {
+  it('should return correct css', () => {
+    expect(getComponentCss()).toMatchSnapshot();
+  });
+});